# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AF Admin R1 Leave Approval API",
    "version": "0.1.0",
    "description": "R1 契约；受控 Schema 数据通过领域校验，不作为动态路由定义编译。请求租户必须验证成员资格。"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/applications/{id}": {
      "get": {
        "operationId": "get_applications_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功包裹；分页 data 为 list 与 total",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess"
                }
              }
            }
          },
          "401": {
            "description": "会话无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "403": {
            "description": "功能无权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "404": {
            "description": "资源不可见或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态、幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "422": {
            "description": "字段或配置校验错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "429": {
            "description": "登录限流",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionToken": []
          }
        ]
      }
    },
    "/applications/{id}/releases": {
      "post": {
        "operationId": "post_applications_id_releases",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功包裹；分页 data 为 list 与 total",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess"
                }
              }
            }
          },
          "401": {
            "description": "会话无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "403": {
            "description": "功能无权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "404": {
            "description": "资源不可见或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态、幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "422": {
            "description": "字段或配置校验错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "429": {
            "description": "登录限流",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Publish"
              }
            }
          }
        }
      }
    },
    "/applications/{id}/activate-release": {
      "post": {
        "operationId": "post_applications_id_activate_release",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功包裹；分页 data 为 list 与 total",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess"
                }
              }
            }
          },
          "401": {
            "description": "会话无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "403": {
            "description": "功能无权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "404": {
            "description": "资源不可见或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态、幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "422": {
            "description": "字段或配置校验错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "429": {
            "description": "登录限流",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivateRelease"
              }
            }
          }
        }
      }
    },
    "/leave-requests": {
      "get": {
        "operationId": "get_leave_requests",
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功包裹；分页 data 为 list 与 total",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess"
                }
              }
            }
          },
          "401": {
            "description": "会话无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "403": {
            "description": "功能无权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "404": {
            "description": "资源不可见或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态、幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "422": {
            "description": "字段或配置校验错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "429": {
            "description": "登录限流",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionToken": []
          }
        ]
      },
      "post": {
        "operationId": "post_leave_requests",
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功包裹；分页 data 为 list 与 total",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess"
                }
              }
            }
          },
          "401": {
            "description": "会话无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "403": {
            "description": "功能无权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "404": {
            "description": "资源不可见或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态、幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "422": {
            "description": "字段或配置校验错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "429": {
            "description": "登录限流",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLeave"
              }
            }
          }
        }
      }
    },
    "/leave-requests/{id}": {
      "get": {
        "operationId": "get_leave_requests_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功包裹；分页 data 为 list 与 total",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess"
                }
              }
            }
          },
          "401": {
            "description": "会话无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "403": {
            "description": "功能无权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "404": {
            "description": "资源不可见或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态、幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "422": {
            "description": "字段或配置校验错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "429": {
            "description": "登录限流",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionToken": []
          }
        ]
      },
      "patch": {
        "operationId": "patch_leave_requests_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功包裹；分页 data 为 list 与 total",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess"
                }
              }
            }
          },
          "401": {
            "description": "会话无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "403": {
            "description": "功能无权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "404": {
            "description": "资源不可见或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态、幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "422": {
            "description": "字段或配置校验错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "429": {
            "description": "登录限流",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLeave"
              }
            }
          }
        }
      }
    },
    "/leave-requests/{id}/submit": {
      "post": {
        "operationId": "post_leave_requests_id_submit",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功包裹；分页 data 为 list 与 total",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccess"
                }
              }
            }
          },
          "401": {
            "description": "会话无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "403": {
            "description": "功能无权",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "404": {
            "description": "资源不可见或不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态、幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "422": {
            "description": "字段或配置校验错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          },
          "429": {
            "description": "登录限流",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiFailure"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Command"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Access-Token"
      }
    },
    "schemas": {
...
      "CreateLeave": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "applicationReleaseId",
          "fields"
        ],
        "properties": {
          "applicationReleaseId": {
            "type": "string"
          },
          "fields": {
            "$ref": "#/components/schemas/LeaveFields"
          },
          "previousRequestId": {
            "type": "string"
          }
        }
      },
      "UpdateLeave": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "expectedRevision",
          "fields"
        ],
        "properties": {
          "expectedRevision": {
            "type": "integer",
            "minimum": 1
          },
          "fields": {
            "$ref": "#/components/schemas/LeaveFields"
          },
          "applicationReleaseId": {
            "type": "string"
          }
        }
      },
      "Publish": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "formDraftId",
          "workflowDraftId",
          "formRevision",
          "workflowRevision",
          "expectedRevision"
        ],
        "properties": {
          "formDraftId": {
            "type": "string"
          },
          "workflowDraftId": {
            "type": "string"
          },
          "formRevision": {
            "type": "integer",
            "minimum": 1
          },
          "workflowRevision": {
            "type": "integer",
            "minimum": 1
          },
          "expectedRevision": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "ActivateRelease": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "releaseId",
          "expectedRevision"
        ],
        "properties": {
          "releaseId": {
            "type": "string"
          },
          "expectedRevision": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "ApiSuccess": {
        "type": "object",
        "required": [
          "code",
          "data"
        ],
        "properties": {
          "code": {
            "const": 20000
          },
          "data": {},
          "traceId": {
            "type": "string"
          }
        }
      },
      "ApiFailure": {
        "type": "object",
        "required": [
          "code",
          "businessCode",
          "message",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "integer"
          },
          "businessCode": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "traceId": {
            "type": "string"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "data": {
            "type": "null"
          }
        }
      },
...
    }
  }
}